+++ trunk/procel/test/LogicaNegocio/VerticeTest.java
@@ -98,7 +98,7 @@ public class VerticeTest {
         System.out.println("getSubindice");
         Vertice instance = new Vertice();
         int expResult = 0;
-        int result = instance.getSubindice();
+        long result = instance.getSubindice();
         assertEquals(expResult, result);
         // TODO review the generated test code and remove the default call to fail.
         fail("The test case is a prototype.");
